requirements.txt
numpy~=1.16.5
pandas~=0.25.1
matplotlib~=3.1.1
seaborn~=0.9.0
python-dateutil~=2.8.0
gspread~=3.6.0
tabula-py~=1.4.1